% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Human-PfMOI.R
\name{human_pfmoi_conpars}
\alias{human_pfmoi_conpars}
\title{Human-PfMOI: Individual Constructor Parameters}
\usage{
human_pfmoi_conpars(id, home_patch_id, trip_duration, trip_frequency,
  bweight, age, moi, chx, bite_algorithm, bite_disp = NaN)
}
\arguments{
\item{id}{integer id of this person}

\item{home_patch_id}{the id of their home patch}

\item{trip_duration}{numeric value; average durarion of trips away from home}

\item{trip_frequency}{numeric value; average rate at which they make trips away from home}

\item{bweight}{numeric value; biting weight to this person (relative to others in a patch; should have mean 1)}

\item{age}{numeric age (not currently used)}

\item{moi}{initial multiplicity of infection (integer)}

\item{chx}{initial chemoprotection (boolean)}

\item{bite_algorithm}{0 for poisson, 1 for negative binomial}

\item{bite_disp}{overdispersion parameter for negative binomial biting (NULL for Poisson)}
}
\description{
Generates a single named list with parameters to construct a single
'human_pfsi' object (see inst/include/Human-PfSI.hpp for constructor parameters).
Be aware that all 'id' related fields should be considered 0-indexed (especially the home patch).
}
