% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patch.R
\name{patches_header}
\alias{patches_header}
\title{Patch: Make Logging Headers}
\usage{
patches_header()
}
\description{
Generates a character vector which will be used as the headers of the .csv output will be logged to
}
